Move body buffer copy from before phyics to before render call. This allows it to be skipped when drawing is disabled while still providing a double-buffer benefit.
Refactored rendering method with a task, loop and wait handles. Reduces the constant allocation/overhead of creating a new task every frame.

Bumped up cull distance.

+++ NBodies/UI/Display.cs
@@ -729,7 +729,7 @@ namespace NBodies.UI
 
         private void PressureScaleUpDown_ValueChanged(object sender, EventArgs e)
         {
-            RenderBase.StyleScaleMax = (float)PressureScaleUpDown.Value;
+            RenderBase.StyleScaleMax = (int)PressureScaleUpDown.Value;
         }
 
         private void CenterOnMassButton_Click(object sender, EventArgs e)
